api
implementacion del openai y news api
