+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "eslint": "8.40.0",
     "eslint-config-next": "13.4.1",
+    "newsapi": "^2.4.1",
     "next": "13.4.1",
     "openai": "^3.2.1",
     "react": "18.2.0",
